move knexfile
