+++ data/dbConfig.js
@@ -1,5 +1,5 @@
 const knex = require("knex");
-const knexConfig = require("../knexfile.js");
+const knexConfig = require("./knexfile.js");
 const environment = process.env.NODE_ENV || "development";
 
 module.exports = knex(knexConfig[environment]);
